move pgsql-specific RETURNING out of common
