+++ src/Pgsql/ReturningInterface.php
@@ -6,7 +6,7 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\SqlQuery\Common;
+namespace Aura\SqlQuery\Pgsql;
 
 /**
  *
