+++ src/Pgsql/Update.php
@@ -17,28 +17,13 @@ use Aura\SqlQuery\Common;
  * @package Aura.SqlQuery
  *
  */
-class Update extends Common\Update implements Common\ReturningInterface
+class Update extends Common\Update implements ReturningInterface
 {
+    use ReturningTrait;
+
     protected function build()
     {
         return parent::build()
             . $this->builder->buildReturning($this->returning);
-    }
-
-    /**
-     *
-     * Adds returning columns to the query.
-     *
-     * Multiple calls to returning() will append to the list of columns, not
-     * overwrite the previous columns.
-     *
-     * @param array $cols The column(s) to add to the query.
-     *
-     * @return $this
-     *
-     */
-    public function returning(array $cols)
-    {
-        return $this->addReturning($cols);
     }
 }
